src/main/java/com/bci/exercise/user/dto/UsuarioDTO.java
package com.bci.exercise.user.dto;

import java.util.List;

public class UsuarioDTO {

    private String name;
    private String email;
    private String password;
    private List<PhoneDTO> phones;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
    public List<PhoneDTO> getPhones() {
        return phones;
    }
    public void setPhones(List<PhoneDTO> phones) {
        this.phones = phones;
    }
    public UsuarioDTO(String name, String email, String password, List<PhoneDTO> phones) {
        this.name = name;
        this.email = email;
        this.password = password;
        this.phones = phones;
    }
    public UsuarioDTO() {

    }
}
